rename GeneralError component to GeneralErrorModal and move to elements
