+++ src/elements/GeneralErrorModal/GeneralErrorModal.styles.ts
